# openapi.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/compliance_definitions/{name}/organization/{orgId}": {
            "post": {
                "operationId": "ComplianceController_createOrgCompliance",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "required": true,
                        "in": "path",
                        "example": "{{test_compliance_definition_name}}",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            },
            "delete": {
                "operationId": "ComplianceController_deactivateCompliance",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "name",
                        "required": true,
                        "in": "path",
                        "example": "{{test_compliance_definition_name}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            }
        },
        "/compliance_definitions/organization/{orgId}": {
            "get": {
                "operationId": "ComplianceController_getOrgComp",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns all activated compliances for an organization"
                    }
                },
                "tags": [
                    "Compliance Definitions"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid"
                        ]
                    }
                ]
            }
        },
...
        "/organizations": {
            "post": {
                "operationId": "CreateOrganization",
                "summary": "Create Organization",
                "parameters": [
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "display_name": "Test API Organization {{$randomAlphaNumeric}}{{$randomAlphaNumeric}}",
                                    "branding": {},
                                    "city": "Minneapolis",
                                    "email": "info@example.com",
                                    "phone": "18005551212",
                                    "state": "MN",
                                    "street_address": "404 Washington Avenue North",
                                    "zip": "55401"
                                },
                                "$ref": "#/components/schemas/CreateOrganizationDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            },
            "get": {
                "operationId": "GetOrganizations",
                "summary": "Get Organizations",
                "parameters": [
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "filter",
                        "required": false,
                        "in": "query",
                        "deprecated": true,
                        "description": "Optional: provide either tha name or the id to filter results",
                        "example": {
                            "name": "{{test_company_name}}",
                            "id": "{{test_company_id}}"
                        },
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            }
        },
        "/organizations/{orgId}/invitation": {
            "post": {
                "operationId": "InviteUser",
                "summary": "Invite User",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "description": "Id of the Organization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "suppressEmail",
                        "required": false,
                        "in": "query",
                        "description": "Suppress Email",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": {
                                    "user_email": "test.owner@example.com",
                                    "inviter_name": "Cold Climate",
                                    "roleId": "company:owner"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            }
        },
        "/organizations/{orgId}/members": {
            "get": {
                "operationId": "GetOrganizationMembers",
                "summary": "Get Organization Members",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "description": "Id of the Organization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            },
            "delete": {
                "operationId": "DeleteOrgMembers",
                "summary": "Delete Members",
                "description": "Deletes specified member ids from organization",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "description": "Id of the Organization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "example": {
                                    "members": [
                                        "{{example_member_id}}"
                                    ]
                                },
                                "description": "Array of member ids"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            }
        },
        "/organizations/{orgId}": {
            "get": {
                "operationId": "GetOrganization",
                "summary": "Get Organization",
                "description": "Get Organization by Name or ID",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "description": "Id of the Organization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            },
            "delete": {
                "operationId": "DeleteOrg",
                "summary": "Delete Organization",
                "description": "Deletes specified organization",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "description": "Id of the Organization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            }
        },
        "/organizations/{orgId}/roles/{roleName}/members/{userId}": {
            "put": {
                "operationId": "AddMemberByID",
                "summary": "Add Member By ID",
                "description": "Add an existing member to the organization",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "description": "Id of the Organization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "roleName",
                        "required": true,
                        "in": "path",
                        "description": "Role Name e.g. `company:owner`",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_owner_id}}",
                        "description": "User Id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            },
            "patch": {
                "operationId": "AddOrgMemberToRole",
                "summary": "Add Member To Role",
                "description": "Adds org member to a role",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_organization_id}}",
                        "description": "Id of the Organization",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "roleName",
                        "required": true,
                        "in": "path",
                        "description": "Role Name e.g. `company:owner`",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "userId",
                        "required": true,
                        "in": "path",
                        "example": "{{test_owner_id}}",
                        "description": "User Id",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "bpc",
                        "required": false,
                        "in": "query",
                        "description": "Bypass Cache",
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            }
        },
...
        "/organizations/{orgId}/files": {
            "post": {
                "operationId": "OrganizationController_uploadFile",
                "parameters": [
                    {
                        "name": "orgId",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Organizations"
                ],
                "security": [
                    {
                        "oauth2": [
                            "openid",
                            "email",
                            "profile"
                        ]
                    }
                ]
            }
        }
    },
    "info": {
        "title": "V1 REST API",
        "description": "🌱 This is a restful api for the cold platform 🌱 ",
        "version": "0.0.0",
        "contact": {},
        "termsOfService": "http://www.coldclimate.com/tos"
    },
...
    "servers": [
        {
            "url": "https://api.coldclimate.com/v1",
            "description": "Production Server"
        },
        {
            "url": "https://api.coldclimate.online/v1",
            "description": "Staging Server"
        }
    ],
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "implicit": {
                        "authorizationUrl": "https://dev-6qt527e13qyo4ls6.us.auth0.com/authorize?audience=https://api.coldclimate.wtf/",
                        "scopes": {
                            "all": "openid profile email"
                        }
                    }
                }
            }
        },
        "schemas": {
...
            "CreateOrganizationDto": {
                "type": "object",
                "properties": {}
            }
        }
    }
}